aix gcc detection and shared library soname support
1) Match "powerpc-ibm-aix5.3.0.0-gcc" as gcc too.

2) Creating shared library on aix with full soname support, following
http://lists.gnu.org/archive/html/libtool/2011-01/msg00023.html
http://lists.gnu.org/archive/html/libtool/2011-02/msg00000.html
Once there is an agreement on that, this patch basically is designed
for upstream perl to be applied when their soname would change anyway.
However, there is 'strip' being used, and I don't know yet what to
use instead in cross-aix setups with recent binutils.

Patch-Name: gentoo/aix_soname.diff